+++ changelog.md
@@ -16,6 +16,7 @@
 ### 👌 Minor Improvements:
 
 - 👌🎨 Wrap model section in result markdown in details tag for notebooks (#1098)
+- 👌 Allow more natural column names in pandas parameters file reading (#1174)
 
 ### 🩹 Bug fixes